+++ toyllm/gpt2/__init__.py
@@ -1,7 +1,7 @@
 from toyllm.gpt2.tokenizer import gpt2_tokenizer
 
 from .config import GPTModelSize
-from .generate import TextGenerator
+from .generate import GptTextGenerator
 from .gpt import GPTModel
 
-__all__ = ["GPTModel", "TextGenerator", "GPTModelSize", "gpt2_tokenizer"]
+__all__ = ["GPTModel", "GptTextGenerator", "GPTModelSize", "gpt2_tokenizer"]
